Item prices come from catalog
This change feels way too big. I'm also not sure about the name of the
new test (same as commit title). I figured I would introduce the catalog
feature by writing a new test that injects the catalog and then does the
'sell one item' flow. I would keep the original tests in place to make
sure I don't break them, only changing/rewriting them once I've gotten
the new test to pass.

Because I introduce a constructor parameter for Checkout, I ended up
having to touch every single test. I tried just giving it a default
parameter of `emptyMap()`, but of course the tests broke because I
changed the implementation of Checkout.total. I extracted the catalog
and the Checkout instance from the new test to properties and deleted
the lines in the other tests that create their own instances of
Checkout.

I originally was just going to put the price for "ipd" in the catalog
because that's the only item I scan in the new test; to accommodate the
other tests, I also added "mbp".