Only deduct price of 1 vga when buying 1 mbp and 2 vga
Turns out this case is already works with the existing implementation.
Keep the test or throw it away? I think I'll keep it for now because
it's definitely important that it behaves this way.

I suppose I will keep trying to triangulate the implementation with
examples, but I'm pondering if it might make sense to replace such tests
with a property test. The property being that for each pair of (mbp,vga)
in the cart, the price of one vga is deducted from the total.

As I write the above, an implementation popped into my thoughts:
 - make a list of every mbp in the cart
 - make a list of every vga in the cart
 - zip the two lists into a list of (mbp,vga) pairs
 - for each pair, deduct the price of one vga
   (or simply deduct pairs.length*priceOfVga)

The exercise doesn't have the concept of order lines (at least not in
the public API of Checkout), so I could get away with simply:
  total -= min(count("mbp"), count("vga")) * priceOfVga
But simply subtracting the discount from the total is going to be a
problem later when the (imaginary) customer comes back next week and
says "oh, of course we need to be able to print a receipt with the
order lines on it. I just assumed you knew that."

I am reminded of one of the episodes of jbrains' WBITDD course[0] in
which he discusses a design decision exactly like this one: keeping
track of a running total vs. keeping track of items scanned. Running
total is "easier" (now), but items scanned is "simpler". Scanned items
is easier to follow in a debugger; running total is a computation, not
data.

[0]: https://testdrivendevelopment.training/